chore: Rename chart prototype file
